+++ libbdk-hal/bdk-config.h
@@ -26,7 +26,7 @@ static inline uint64_t bdk_config_get(bdk_config_t cfg, uint64_t default_value)
 {
     extern uint64_t __bdk_config_get_slow(bdk_config_t cfg, uint64_t default_value);
     extern uint8_t __bdk_config_is_valid[__BDK_CONFIG_END];
-    extern uint8_t __bdk_config_cache[__BDK_CONFIG_END];
+    extern uint64_t __bdk_config_cache[__BDK_CONFIG_END];
 
     if (bdk_likely(__bdk_config_is_valid[cfg]))
         return __bdk_config_cache[cfg];
